feat(ProgressBar): add breakppoint
- Verify that the ProgressBar slows down on breakpoints and it
lunches forwards to current percentage after it is gone from breakpoint
(the animation can be adjusted depending on taste)
