+++ src/progress_bar_exercise/ProgressBarExercise.js
@@ -1,7 +1,8 @@
-import React, { useCallback, useEffect, useRef, useState } from "react";
+import React, { useState } from "react";
 import Exercise from "../exercise/Exercise";
-import { ProgressBar } from "../shared-components/ProgressBar";
-import { WarningBtn, DefaultBtn } from "../shared-components/Button";
+import { ProgressBarV1 } from "../shared-components/ProgressBarV1";
+import { ProgressBarV2 } from "../shared-components/ProgressBarV2";
+import { DefaultBtn } from "../shared-components/Button";
 
 const ProgressBarExercise = () => {
   return (
@@ -19,80 +20,19 @@ export default ProgressBarExercise;
 
 // ----------------------------------------------------------------------------------
 
-const REQUEST_DURATION = 15000;
-const STOP_PERCENTAGE = 90;
-const PROGRESS_BAR_FADEOUT = 3000;
-
 export const Solution = () => {
-  const [progressBarState, setProgressBarState] = useState({
-    percentage: 0,
-    startTime: null,
-  });
+  const [solution, setSolution] = useState(1);
 
-  const requestRef = useRef();
-  const previousTimeRef = useRef();
-  const fadetimeout = useRef();
-  const isProgressVisible = progressBarState.percentage > 0;
-  const isProgressFinished = progressBarState.percentage === 100;
-
-  const animate = useCallback((time) => {
-    if (
-      previousTimeRef.current !== undefined &&
-      progressBarState.percentage < STOP_PERCENTAGE
-    ) {
-      const timePassed = Date.now() - progressBarState.startTime;
-
-      setProgressBarState({
-        ...progressBarState,
-        percentage: (timePassed * 100) / REQUEST_DURATION,
-      });
-    }
-    previousTimeRef.current = time;
-    requestRef.current = requestAnimationFrame(animate);
-  }, [progressBarState]);
-
-  useEffect(() => {
-    if (isProgressVisible && !isProgressFinished) {
-      clearTimeout(fadetimeout.current)
-      requestRef.current = requestAnimationFrame(animate);
-    }
-    return () => cancelAnimationFrame(requestRef.current);
-  }, [isProgressVisible, isProgressFinished, requestRef, animate]);
-
-  const startProgressBar = () =>
-    setProgressBarState({
-      percentage: 0.1,
-      startTime: Date.now(),
-    });
-
-  const stopProgressBar = () => {
-    setProgressBarState({
-      percentage: 100,
-      startTime: null,
-    });
-    fadetimeout.current = setTimeout(resetProgressBar, PROGRESS_BAR_FADEOUT)
-  };
-
-  const resetProgressBar = () =>
-    setProgressBarState({
-      ...progressBarState,
-      percentage: 0,
+  const toggleSolution = () =>
+    setSolution((prevSolution) => {
+      if (prevSolution === 1) return 2;
+      if (prevSolution === 2) return 1;
     });
 
   return (
     <>
-      {isProgressVisible && (
-        <ProgressBar currentPercentage={progressBarState.percentage} />
-      )}
-      <DefaultBtn
-        title={
-          isProgressVisible && !isProgressFinished
-            ? "Loading..."
-            : "Start Request"
-        }
-        onClick={startProgressBar}
-      />
-      <WarningBtn title="Finish Request" onClick={stopProgressBar} />
+      <DefaultBtn title="Toggle Solution" onClick={toggleSolution} />
+      {solution === 1 ? <ProgressBarV1 /> : <ProgressBarV2 />}
     </>
   );
 };
